embeds: - move embed receipts into ./src/embed-receipts.js - support automatically setting noCache and noWebWorker for older browsers (eg. safari < 16) - only autohide embed when 'hideOffscreen' attribute is set
verification info:
- show 'not available' if no verification info both on info page and receipt
