+++ src/coll.js
@@ -21,17 +21,10 @@ import fasLeft from "@fortawesome/fontawesome-free/svgs/solid/arrow-left.svg";
 import fasRight from "@fortawesome/fontawesome-free/svgs/solid/arrow-right.svg";
 import fasMenuV from "@fortawesome/fontawesome-free/svgs/solid/ellipsis-v.svg";
 
-//import fasGlobe from "@fortawesome/fontawesome-free/svgs/solid/globe.svg";
-//import fasAngleDoubleDown from "@fortawesome/fontawesome-free/svgs/solid/angle-double-down.svg";
-import btGlobe from "../assets/globe.svg";
-import btAngleDoubleDown from "../assets/chevron-double-down.svg";
-import btAngleDoubleUp from "../assets/chevron-double-up.svg";
-
-import fabGithub from "@fortawesome/fontawesome-free/svgs/brands/github.svg";
-
 import fasAngleLeft from "@fortawesome/fontawesome-free/svgs/solid/angle-left.svg";
 import fasAngleRight from "@fortawesome/fontawesome-free/svgs/solid/angle-right.svg";
 
+import { RWPEmbedReceipt } from "./embed-receipt.js";
 import Split from "split.js";
 
 
@@ -142,7 +135,7 @@ class Coll extends LitElement
       this.isFullscreen = !!document.fullscreenElement;
     });
 
-    if (this.embed) {
+    if (this.embed && this.loadInfo && this.loadInfo.hideOffscreen) {
       this.observer = new IntersectionObserver((entries/*, observer*/) => {
         this.isVisible = entries[0].isIntersecting;
       });
@@ -486,129 +479,12 @@ class Coll extends LitElement
       flex: auto;
     }
 
-    #embed-dropdown {
-      max-height: calc(100vh - 50px);
-      padding-top: 0;
-      margin-top: -0.5rem;
-      display: block;
-      z-index: 1;
-      pointer-events: none;
-      transition: all .3s linear;
-      transform-origin: left top;
-      transform: scaleY(0);
-      transition: all 300ms cubic-bezier(0.15, 0, 0.1, 1);
-      filter: drop-shadow(0px 8px 4px rgba(0, 0, 0, 0.15));
-    }
-
-    .dropdown.is-active #embed-dropdown {
-      transform: scaleY(1);
-    }
-
-    .embed-info-container {
-      width: 100%;
-      display: flex !important;
-      justify-content: center;
-    }
-
-    button.embed-info {
-      padding: 0;
-      background-color: white;
-      justify-content: space-between;
-      max-width: 40rem;
-      width: calc(100% - 1rem);
-      height: 42px;
-      border-color: #D1D5DA;
-      border-width: 1px;
-      border-style: solid;
-      border-radius: 999px;
+    rwp-embed-receipt {
+      flex-direction: column;
       display: flex;
-      align-items: center;
-      text-overflow: ellipsis;
-      filter: drop-shadow(0px 1px 2px rgba(0, 0, 0, 0.15));
-      transition-duration: 50ms;
-      transition-timing-function: ease-out;
-      cursor: pointer;
-      z-index: 2
     }
 
-    button.embed-info:active {
-      color: initial;
-    }
-
-    button.embed-info:hover {
-      filter: drop-shadow(0px 2px 4px rgba(0, 0, 0, 0.2));
-      transform: scale(1.01);
-    }
-
-    button.embed-info:hover:active {
-      transform: translateY(0.25rem);
-    }
-
-    .embed-info-buttontext {
-      white-space: nowrap;
-      overflow: hidden;
-      text-overflow: ellipsis;  
-      flex-grow: 1;
-      text-align: start;
-    }
-
-    .embed-info-drop {
-      font-size: 14px;
-      padding: 1rem;
-      padding-top: 2rem;
-      max-width: 38rem;
-      max-height: 42rem;
-      width: calc(100% - 2rem);
-      border-top-right-radius: 0px;
-      border-top-left-radius: 0px;
-      pointer-events: auto;
-      overflow-y: auto;
-    }
-
-    .embed-info-drop > p {
-      font-size: 14px;
-      color: black;
-    }
-
-    .embed-info-drop > h2 {
-      margin-bottom: 0.25rem;
-      font-size: 16px;
-      font-weight: bold;
-      text-transform: none;
-      letter-spacing: 0;
-      color: #24292E;
-    }
-
-    .embed-info-drop-statscontainer > h3 {
-      font-size: 12px;
-      color: #394146;
-    }
-
-    .embed-info-drop-statscontainer > p {
-      font-size: 14px;
-      color: black;
-    }
-
-    .embed-info-drop a {
-      word-break: break-all;
-    }
-
-    .embed-info-drop .show-hash {
-      word-break: break-all;
-      font-family: monospace;
-    }
-
-    .embed-globe {
-      margin: 0.25rem;
-      padding: 7px;
-      background-color: #0366D6;
-      border-radius: 9999px;
-      color: white;
-      border-width: 1px;
-      border-color: #D1D5DA;
-      border-style: solid;
-      line-height: 0.5em;
-    }
+    ${RWPEmbedReceipt.embedStyles}
 
     ${Coll.replayBarStyles}`;
   }
@@ -1021,87 +897,13 @@ class Coll extends LitElement
       return "";
     }
 
-    let {numValid, numInvalid, domain, certFingerprint, datapackageHash, software} = this.collInfo.verify || {};
-    numValid = numValid || 0;
-    numInvalid = numInvalid || 0;
-
-    const sourceUrl = this.collInfo.sourceUrl;
-
-    const certFingerprintUrl = certFingerprint ? `https://search.censys.io/certificates/${certFingerprint}` : "";
-
-    const dateStr = tsToDate(this.ts).toLocaleString();
-
-    return html`
-    <div class="dropdown mb-4 ${this.embedDropdownActive ? "is-active" : ""}">
-      <div class="dropdown-trigger embed-info-container">
-        <button class="embed-info is-small is-rounded mt-4" aria-haspopup="true" aria-controls="embed-dropdown" @click="${this.onEmbedDrop}">
-          <fa-icon class="menu-logo mr-2 embed-globe" size="1rem" aria-hidden="true" .svg=${btGlobe}></fa-icon>
-          <span class="embed-info-buttontext">
-            This embed is part of a web archive. Click here to learn more.
-          </span>
-          <span class="icon is-small mr-4 ml-2">
-            <fa-icon title="Toggle" .svg="${this.embedDropdownActive ? btAngleDoubleUp : btAngleDoubleDown}" aria-hidden="true"></fa-icon>
-          </span>
-        </button>
-      </div>
-      <div class="dropdown-menu embed-info-container" id="embed-dropdown" role="menu">
-        <div class="dropdown-content embed-info-drop">
-          <h2>This Embed Won't Go Away</h2>
-          <p class="mb-4">
-          The content shown here is loaded from a web archive. Even if the original page
-          goes offline or is changed, the content below will remain unchanged.
-          </p>
-          <hr class="dropdown-divider">
-          <h2 mt-4">Archive Information</h2>
-          <div class="embed-info-drop-statscontainer mb-4">
-            <h3>Original URL:</h3>
-            <p><a target="_blank" href="${this.tabData.url}">${this.tabData.url}</a></p>
-            <h3 class="mt-2">Archived On:</h3>
-            <p>${dateStr}</p>
-            ${domain ? html`
-            <h3 class="mt-2">Observed By:</h3>
-            <p>${domain}</p>
-            ${certFingerprintUrl ?
-    html`<p><a target="_blank" href="${certFingerprintUrl}">View Certificate</a></p>` : ""}
-            ` : software ? html`
-            <h3 class="mt-2">Created With:</h3>
-            <p>${software}</p>` : ""}
-            <h3 class="mt-2">Validation:</h3>
-            ${numValid > 0 || numInvalid > 0 ? html`
-            <p>${numValid} hashes verified${numInvalid ? html`, ${numInvalid} invalid` : ""}</p>` : html`
-            <p>Not Available</p>
-            `}
-            <h3 class="mt-2">Package Hash:</h3>
-            <p class="show-hash">${datapackageHash}</p>
-          </div>
-          ${sourceUrl ? html`
-          <hr class="dropdown-divider">
-          <h2 mt-4">Download Archive</h2>
-          <a target="_blank" href="${sourceUrl}">${sourceUrl}</a>
-          <p class="mt-2">This archive can be viewed directly in your browser by loading it via <a target="_blank" href="https://replayweb.page">replayweb.page</a></p>
-          ` : ""}
-          <p class="is-size-7 is-flex is-justify-content-space-between" style="margin-top: 40px">
-            <span>
-              <a class="has-text-black" target="_blank" href="https://github.com/webrecorder/replayweb.page">
-                <fa-icon class="menu-logo mr-1" size="1.0rem" aria-hidden="true" .svg=${this.appLogo}></fa-icon>
-                Powered by ReplayWeb.page
-              </a>
-            </span>
-            <span>
-              <a class="has-text-black" target="_blank" href="https://github.com/webrecorder/replayweb.page">Source Code
-                <fa-icon class="menu-logo ml-1" size="1.0rem" aria-hidden="true" .svg=${fabGithub}></fa-icon>
-              </a>
-            </span>
-          </p>
-        </div>
-      </div>
-    </div>
-    `;
-  }
-
-  onEmbedDrop(event) {
-    event.stopPropagation();
-    this.embedDropdownActive = !this.embedDropdownActive;
+    return html`<rwp-embed-receipt
+            .collInfo=${this.collInfo}
+            url=${this.url}
+            ts=${this.ts}
+            .appLogo=${this.appLogo}
+            >
+            </rwp-embed-receipt>`;
   }
 
   dragStart() {
